+++ vhost-user-backend/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - [[#374]](https://github.com/rust-vmm/vhost/pull/374) Call backend event handler on vring enabled
+- [[#377]](https://github.com/rust-vmm/vhost/pull/377) Add a `stop_vring()` hook for draining requests before `GET_VRING_BASE` returns.
 
 ### Changed
 ### Deprecated